[WIP] Merge API

+++ test/github.js
@@ -3,21 +3,26 @@ var _ = require('lodash');
 
 var Octocat = require('octocat');
 var repofs = require('../');
+var conflicter = require('../lib/conflicter');
 
-var GH_REPO = process.env.GITHUB_REPO;
-var GH_TOKEN = process.env.GITHUB_TOKEN;
+var MODE = process.env.REPOFS_MODE || 'github';
+var REPO = process.env.REPOFS_REPO;
+var HOST = process.env.REPOFS_HOST;
+var TOKEN = process.env.REPOFS_TOKEN;
 
-if (!GH_TOKEN || !GH_REPO) throw new Error('Testing require github configuration');
+if (!REPO || !HOST) throw new Error('Testing require github/uhub configuration');
 
 var client = new Octocat({
-    token: GH_TOKEN
+    token: TOKEN
 });
 
-describe('GitHub Driver', function() {
+describe('GitHub/uHub Driver', function() {
     var commit;
+    var repo = client.repo(REPO);
     var fs = repofs({
-        repository: GH_REPO,
-        token: GH_TOKEN,
+        repository: REPO,
+        token: TOKEN,
+        host: HOST,
         committer: {
             name: 'John Doe',
             email: 'johndoe@gmail.com'
@@ -26,7 +31,7 @@ describe('GitHub Driver', function() {
 
     // Setup base repo
     before(function() {
-        var repo = client.repo(GH_REPO);
+        if (MODE != 'github') return;
 
         return repo.destroy()
         .fail(function(err) {
@@ -34,7 +39,7 @@ describe('GitHub Driver', function() {
         })
         .then(function() {
             return client.createRepo({
-                name: _.last(GH_REPO.split('/')),
+                name: _.last(REPO.split('/')),
                 auto_init: true
             });
         });
@@ -42,7 +47,7 @@ describe('GitHub Driver', function() {
 
     // Destroy repository after tests
     after(function() {
-        var repo = client.repo(GH_REPO);
+        if (MODE != 'github') return;
         return repo.destroy();
     });
 
@@ -69,11 +74,11 @@ describe('GitHub Driver', function() {
 
     describe('fs.read', function() {
         it('should correctly read from master', function() {
-            return fs.read('README.md').should.be.fulfilled;
+            return fs.read('README.md').should.be.fulfilled();
         });
 
         it('should fail for non-existant file', function() {
-            return fs.read('error-repofs.js').should.be.rejected;
+            return fs.read('error-repofs.js').should.be.rejected();
         });
     });
 
@@ -101,13 +106,13 @@ describe('GitHub Driver', function() {
 
     describe('fs.write', function() {
         it('should fail to write non-existing file', function() {
-            return fs.write('README_nonexistant.md', 'test').should.be.rejected;
+            return fs.write('README_nonexistant.md', 'test').should.be.rejected();
         });
     });
 
     describe('fs.create', function() {
         it('should fail to create existing file', function() {
-            return fs.create('README.md', 'test').should.be.rejected;
+            return fs.create('README.md', 'test').should.be.rejected();
         });
     });
 
@@ -144,22 +149,22 @@ describe('GitHub Driver', function() {
         });
 
         it('should create a file', function() {
-            return fs.create('TEST.md', 'test create').should.be.fulfilled;
+            return fs.create('TEST.md', 'test create').should.be.fulfilled();
         });
 
         it('should signal that file has been created', function() {
-            var changes = fs.listChanges();
+            var changes = fs.working.listChanges();
             changes.should.have.property('TEST.md');
             changes['TEST.md'].type.should.equal('create');
             changes.should.not.have.property('README.md');
         });
 
         it('should edit a file', function() {
-            return fs.write('README.md', 'test edit').should.be.fulfilled;
+            return fs.write('README.md', 'test edit').should.be.fulfilled();
         });
 
         it('should signal that file has been edited', function() {
-            var changes = fs.listChanges();
+            var changes = fs.working.listChanges();
             changes.should.have.property('TEST.md');
             changes.should.have.property('README.md');
             changes['README.md'].type.should.equal('update');
@@ -168,12 +173,202 @@ describe('GitHub Driver', function() {
         it('should correctly commit change', function() {
             return fs.commit({
                 message: 'Test commit'
-            }).should.be.fulfilled;
+            }).should.be.fulfilled();
         });
 
         it('should have cleared changes', function() {
-            var changes = fs.listChanges();
+            var changes = fs.working.listChanges();
             _.size(changes).should.equal(0);
+        });
+    });
+
+    describe('Handling branches', function() {
+        var dir = 'branchdir/';
+        var appendContent = 'Both will write without conflict';
+
+        // Structure of the directory for merging tests
+        // branchdir
+        // ├── appendfile	Both modified without conflict
+        // ├── branchfile	Created by branch only
+        // ├── conflictfile	Both modified with conflict
+        // └── masterfile	Created by master only
+
+        before(function() {
+            return fs.checkout('master')
+                .then(function () {
+                    return fs.update(dir + 'masterfile', 'Only master writes here');
+                })
+                .then(function () {
+                    return fs.update(dir + 'appendfile', appendContent);
+                })
+                .then(function () {
+                    return fs.update(dir + 'conflictfile', 'Both will write with conflict');
+                })
+                .then(function () {
+                    return fs.commit({ message: 'Files setup'});
+                })
+                // TODO branch creations could be done in parallel
+                .then(function () {
+                    // Branch with conflicts
+                    return fs.createBranch('conflict');
+                })
+                .then(function () {
+                    // Branch without conflict
+                    return fs.createBranch('no-conflict');
+                })
+                .then(function () {
+                    return fs.update(dir + 'conflictfile', 'Both will write with MASTER conflict');
+                })
+                .then(function () {
+                    return fs.commit({ message: 'Master makes conflict'});
+                })
+                .then(function () {
+                    return fs.checkout('no-conflict');
+                })
+                .then(function () {
+                    return fs.update(dir + 'branchfile', 'Only the branch is writing');
+                })
+                .then(function () {
+                    return fs.update(dir + 'appendfile', appendContent
+                                     + ' Branch safely append here');
+                })
+                .then(function () {
+                    return fs.commit({ message: 'Creates branchfile. Appends to appendfile'});
+                })
+                .then(function () {
+                    return fs.checkout('conflict');
+                })
+                .then(function () {
+                    return fs.update(dir + 'conflictfile', 'Both will write with BRANCH conflict');
+                })
+                .then(function () {
+                    return fs.commit({ message: 'Conflicting write to conflictfile'});
+                })
+                .then(function () {
+                    return fs.checkout('master');
+                });
+        });
+
+        it('should compare identical branches', function() {
+            return fs._detectConflicts('master', 'master')
+                .then(function (r) {
+                    r.status.should.eql(conflicter.REFS.IDENTICAL);
+                    r.base.should.eql('master');
+                    r.head.should.eql('master');
+                });
+        });
+
+        it('should compare diverging branches', function() {
+            return fs._detectConflicts('master', 'conflict')
+                .then(function (r) {
+                    r.status.should.eql(conflicter.REFS.DIVERGED);
+                    r.base.should.eql('master');
+                    r.head.should.eql('conflict');
+                    r.conflicts[dir + 'conflictfile'].should.eql({
+                        base: 'ecf9936de81913a4b292d1adf46a3e2e9b6bae95',
+                        head: '05dd87fed0042c852a51afc389c9a341c5a0b4a2',
+                        path: dir + 'conflictfile',
+                        status: conflicter.FILE.BOTH_MODIFIED
+                    });
+                });
+        });
+
+        it('should automatically merge non-conflicting branches', function() {
+            return fs.createBranch('merge-no-conflict') // = master
+                .then(function () {
+                    return fs.checkout('merge-no-conflict');
+                })
+                .then(function() {
+                    return fs.mergeBranches('no-conflict', 'merge-no-conflict');
+                })
+                .then(function() {
+                    return fs.checkout('master');
+                });
+        });
+
+        it('should raise a conflict event when merging conflicting branches', function() {
+            fs.once('conflicts.resolve.needed', function(conflicts, next) {
+                // Deleting everything is a good way of solving conflicts.
+                // Thinking about making it the default... ;)
+                next(null, {
+                    message: 'Cleaned up',
+                    conflicts: {
+                    }
+                });
+            });
+            return fs.createBranch('merge-conflict-event') // = master
+            .then(function () {
+                return fs.checkout('merge-conflict-event');
+            })
+            .then(function () {
+                return fs.mergeBranches('conflict', 'merge-conflict-event');
+            })
+            .then(function () {
+                return fs.checkout('master');
+            });
+        });
+
+        it('should solve conflicts when merging', function() {
+            fs.once('conflicts.resolve.needed', function(conflicts, next) {
+                next(null, {
+                    message: 'Merge message',
+                    files: {
+                        'branchdir/conflictfile': {
+                            path: dir+'conflictfile',
+                            buffer: 'Merged content' }
+                    }
+                });
+            });
+            return fs.createBranch('merge-conflict') // = master
+            .then(function () {
+                return fs.checkout('merge-conflict');
+            })
+            .then(function () {
+                return fs.mergeBranches('conflict', 'merge-conflict');
+            })
+            .then(function () {
+                return Q.all([
+                    fs.read(dir+'conflictfile'),
+                    fs.read(dir+'masterfile')
+                ]);
+            })
+            .spread(function (conflictContent, masterContent) {
+                masterContent.should.be.eql('Only master writes here');
+                conflictContent.should.be.eql('Merged content');
+            })
+            .then(function () {
+                return fs.checkout('master');
+            });
+        });
+
+        it('should fail if conflicts are not resolved', function() {
+            fs.once('conflicts.resolve.needed', function(conflicts, next) {
+                next(new Error('Failed, because of reasons'));
+            });
+            return fs.createBranch('merge-conflict-fail') // = master
+            .then(function () {
+                return fs.checkout('merge-conflict-fail');
+            })
+            .then(function () {
+                return fs.mergeBranches('conflict', 'merge-conflict-fail');
+            })
+            .should.be.rejected();
+        });
+
+        it('should compare non conflicting-branches branches', function() {
+            // TODO
+        });
+
+        it('should compare non fast-forwardable branches (behind)', function() {
+            // TODO
+        });
+
+        it('should compare non fast-forwardable branches (ahead)', function() {
+            // TODO
+        });
+
+        it('should compare conflicting-branches branches', function() {
+            // TODO
         });
     });
 });
